#144 fix: controller 수정
